# services/auth/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SaintAugustin Auth Service",
        "description": "Authentication, user management, and role assignment for the SaintAugustin worship platform.",
        "contact": {
            "name": "SaintAugustin",
            "email": "admin@saintaugustin.local"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Local development (via API gateway)"
        }
    ],
    "paths": {
...
        "/imports/videopsalm": {
            "post": {
                "tags": [
                    "Imports"
                ],
                "summary": "Import (or preview) songs from a VideoPsalm `.vpagd` archive",
                "description": "Accepts a multipart upload of a `.vpagd` file. With `dry_run=true` the parser returns the song list without writing anything; without it (or with `dry_run=false`), the songs are persisted and the response reports `created` / `skipped` counts. Songs are de-duplicated by case-insensitive (title, songbook) — re-running a previous import skips existing rows rather than overwriting hand-edits. Requires the `admin` role.",
                "operationId": "fe1b2212612cd8082c550160fbba3fee",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "description": "The `.vpagd` archive (a ZIP of Song_*.json files).",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "dry_run": {
                                        "description": "If true, only parse and return preview rows; nothing is persisted.",
                                        "type": "boolean",
                                        "default": false
                                    },
                                    "guids": {
                                        "description": "Optional Guid filter — only songs whose Guid is in the list are imported.",
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Preview (dry run) — list of songs that would be imported, plus a per-songbook count.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "songs": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "songbooks": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "integer"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "Import committed — counts of created/skipped songs.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "created": {
                                            "type": "integer"
                                        },
                                        "skipped": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        },
                                        "songbooks": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Admin role required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "File missing, too large, or unparseable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorResponse": {
                "description": "Generic error response.",
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Invalid credentials."
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "description": "Laravel validation failure (HTTP 422).",
                "required": [
                    "message",
                    "errors"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The email field is required."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "email": [
                                "The email field is required."
                            ]
                        },
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter your Bearer token. Obtain it from POST /api/auth/login."
            }
        }
    },
...
}